initial XBee library
I wrote a lot of code and I have no idea if it works! But here are some
things to note:

First, I wrote some public helper methods (such as sendAtCommand())
which take Strings by reference and then concatenate them into new
Strings. This may lead to heap fragmentation, I really don't know. A
probably safer way to implement this is to have a legitimate transmit
buffer.

Secondly, I wrote a read() method. It returns a userPacket, which is a
struct that contains references to either null values (NULL_USER_PACKET)
or references to parts of a receive buffer (m_rxBuffer). There are some
things that I don't know will work properly:
1. When I set m_rxBuffer = {} to clear the buffer, will the contained
String frameData be deallocated? If so, will its reallocation and
re-reservation in the constructor cause heap fragmentation?
2. What actually happens when I return NULL_USER_PACKET, since I'm not
returning the struct itself by reference, but the struct *contains*
references?

Anyway, I'm pretty proud of this. Even if it doesn't work. At least it
compiles.