shutdown via API mode almost works
The thing that fixed it was adding a frameID field between frameType and
frameData that wasn't documented well.

Had to refactor the code to use char arrays with length instead of Strings
because if a String intentionally contains the null character, it won't
be interpreted properly. I haven't tested since the refactor; that's for
Saturday or Monday.