got response
Pretty sure we're getting responses now that we've disabled the
MonitoredSerial. We think the hardware serial buffers are getting
overflowed, and we're not reading them fast enough. Circumventing
MonitoredSerial increases the speed we are able to read. However, we are
getting garbage on the computer's serial monitor now, just from the
debug messages that the XBee sends.

Preferably, we have a solution where we can monitor/debug all of the
bytes being sent and received, while still avoiding hardware buffer
overflows (if that is in fact the problem, and I don't know what else
would be).