make testStats volatile-qualified
I'm gonna be real with you chief, I don't understand everything that I
did. For example, why do I have to volatile-qualify all of the methods
of TelemetryData? And why, when I write a function that takes a
volatile-qualified TelemetryData, does it automatically have to be a
TelemetryData& instead?

These things kind of make sense but I'm not completely sure about them.
Anyway it works so I'm moving on.