initially set the data if not present
Fixed bug where a few values were being transmitted, then no more were
being transmitted. This is because clear()ing only changes the present
status of each member in TelemetryData; it doesn't touch the data in the
union.

This probably needs to be designed better so that you can't just call
getDouble() when the double is not present; it's just that you can't
throw exceptions on this thing.